Project Development Phase/Sprint 4/requirements.txt
blinker==1.5
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Mail==0.9.1
ibm-db==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.0
python-http-client==3.3.7
PythonMail==0.0.0.4
sendgrid==6.9.7
starkbank-ecdsa==2.2.0
Werkzeug==2.2.2
blinker==1.5
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Mail==0.9.1
ibm-db==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.0
python-http-client==3.3.7
PythonMail==0.0.0.4
sendgrid==6.9.7
starkbank-ecdsa==2.2.0
Werkzeug==2.2.2
